Move onboarding tours logic within a reusable hook.
